lightspeed: move our VueJs components in a dedicated directory
The webviews/lightspeed directory will soon be removed, by moving our
components in a dedicated directory, we make room for the new
components that are coming.
