+++ webviews/lightspeed/src/components/lightspeed/PromptExampleBox.vue
