+++ webviews/lightspeed/src/components/lightspeed/StatusBoxCollectionName.vue
