+++ webviews/lightspeed/src/components/lightspeed/StatusBoxPrompt.vue
